+++ engine/core/platform.h
@@ -6,6 +6,7 @@
 #define PLATFORM_H_
 
 #include <core/imgui/imgui.h>
+#include <stdint.h>
 
 namespace display
 {
@@ -24,11 +25,11 @@ namespace platform
 		virtual void OnInit() = 0;
 		virtual void OnDestroy() = 0;
 		virtual void OnTick(double total_time, float elapsed_time) = 0;
-		virtual void OnSizeChange(size_t width, size_t height, bool minimized) = 0;
+		virtual void OnSizeChange(uint32_t width, uint32_t height, bool minimized) = 0;
 		virtual void OnAddImguiMenu() {};
 	};
 
-	char Run(const char* name, void* param, size_t width, size_t height, Game* game);
+	char Run(const char* name, void* param, uint32_t width, uint32_t height, Game* game);
 }
 
 #endif PLATFORM_H_
